pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "email-elchemy"
version = "0.1.0"
description = "Ingest Gmail threads, extract quotes, and expose APIs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Email Elchemy" }
]
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "redis>=5.0",
  "rq>=1.16",
  "httpx>=0.27",
  "google-api-python-client>=2.141.0",
  "google-auth-oauthlib>=1.2.1",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "openai>=1.40.0",
  "jsonschema>=4.23.0",
  "pdfminer.six>=20240706",
  "python-docx>=1.1.2",
  "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.6",
  "black>=24.8.0",
  "pytest>=8.3",
  "mypy>=1.11.1",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = [
  "E","F","I","B","UP","SIM","ASYNC","A","PL",
]
lint.ignore = [
  "E501",  # handled by black
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"
